Move visit method bodies to DockerfileVisitor
